+++ src/components/organisms/homework/homework.module.scss
@@ -1,6 +1,9 @@
 .form {
   margin-top: 16px;
 }
+.description {
+  margin-bottom: 16px;
+}
 .status {
   margin-top: 16px;
 }
